net/txgbe: rename raptor to sp for Sapphire-specific code
The term "raptor" was originally used as a general designation for
NICs when only Sapphire series existed. With the addition of
Amber-Lite series NICs, we now use "sp" to replace "raptor" for
driver components that are specific to Sapphire NICs, while
retaining the "raptor" name for common flows applicable to all NICs.

Signed-off-by: Zaiyu Wang <zaiyuwang@trustnetic.com>